requirements.txt

Flask==3.0.3
gunicorn==22.0.0
google-cloud-datastore==2.19.0
google-cloud-pubsub==2.21.5
